Emailing script + mail template

+++ app/routes.js
@@ -43,8 +43,12 @@ module.exports = function (app, passport) {
     .get ('/items/:itemId', controllers.api.getItemById)
     .post ('/create', controllers.api.createItem)
     .delete ('/delete/:itemId', controllers.api.deleteItem)
-    .get ('/delete/:itemId', controllers.api.deleteItem)
     .get ('/toggle_status/:itemId', controllers.api.toggleItemStatus);    //toggle status -> if item.active = true, item still needs to be completed, if item.active = false, it has been completed
+
+    if (process.env.NODE_ENV === 'development') {
+      apiRouter
+        .get ('/delete/:itemId', controllers.api.deleteItem);
+    }
 
   app.use ('/api', apiRouter);
